Gnome Health functions and Items
Added three functions to the gnome blueprint. TakeDamage, Heal, and Die. TakeDamage and Heal both take in an int variable as arguments.The functions are still all preliminary and should be added to and improved in the future.

Created a child class of the ItemParent class called HealthPack. Its item type is set to “Consumable” in the constructor